Fix, refactor VBD solver
