+++ newton/solvers/vbd/solver_vbd.py
@@ -21,15 +21,15 @@ from warp.types import float32, matrix
 
 from newton.core.types import override
 from newton.geometry import PARTICLE_FLAG_ACTIVE, Contacts
-from newton.geometry.kernels import (
-    TriMeshCollisionDetector,
-    TriMeshCollisionInfo,
-    triangle_closest_point,
-)
+from newton.geometry.kernels import triangle_closest_point
 from newton.sim import Control, Model, State
 from newton.sim.model import ShapeMaterials
 
 from ..solver import SolverBase
+from .tri_mesh_collision import (
+    TriMeshCollisionDetector,
+    TriMeshCollisionInfo,
+)
 
 # TODO: Grab changes from Warp that has fixed the backward pass
 wp.set_module_options({"enable_backward": False})
@@ -491,52 +491,6 @@ def evaluate_dihedral_angle_based_bending_force_hessian(
 
 
 @wp.func
-def evaluate_ground_contact_force_hessian(
-    particle_pos: wp.vec3,
-    particle_prev_pos: wp.vec3,
-    particle_radius: float,
-    ground_normal: wp.vec3,
-    ground_level: float,
-    soft_contact_ke: float,
-    soft_contact_kd: float,
-    friction_mu: float,
-    friction_epsilon: float,
-    dt: float,
-):
-    penetration_depth = -(wp.dot(ground_normal, particle_pos) + ground_level - particle_radius)
-
-    if penetration_depth > 0:
-        ground_contact_force_norm = penetration_depth * soft_contact_ke
-        ground_contact_force = ground_normal * ground_contact_force_norm
-        ground_contact_hessian = soft_contact_ke * wp.outer(ground_normal, ground_normal)
-
-        dx = particle_pos - particle_prev_pos
-
-        if wp.dot(dx, ground_normal) < 0:
-            damping_hessian = (soft_contact_kd / dt) * ground_contact_hessian
-            ground_contact_hessian = ground_contact_hessian + damping_hessian
-            ground_contact_force = ground_contact_force - damping_hessian * dx
-
-        # friction
-        e0, e1 = build_orthonormal_basis(ground_normal)
-
-        T = mat32(e0[0], e1[0], e0[1], e1[1], e0[2], e1[2])
-
-        relative_translation = dx
-        u = wp.transpose(T) * relative_translation
-        eps_u = friction_epsilon * dt
-
-        friction_force, friction_hessian = compute_friction(friction_mu, ground_contact_force_norm, T, u, eps_u)
-        ground_contact_force = ground_contact_force + friction_force
-        ground_contact_hessian = ground_contact_hessian + friction_hessian
-    else:
-        ground_contact_force = wp.vec3(0.0, 0.0, 0.0)
-        ground_contact_hessian = wp.mat33(0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0)
-
-    return ground_contact_force, ground_contact_hessian
-
-
-@wp.func
 def evaluate_body_particle_contact(
     particle_index: int,
     particle_pos: wp.vec3,
@@ -548,7 +502,7 @@ def evaluate_body_particle_contact(
     friction_epsilon: float,
     particle_radius: wp.array(dtype=float),
     shape_materials: ShapeMaterials,
-    shape_geo: wp.array(dtype=int),
+    shape_body: wp.array(dtype=int),
     body_q: wp.array(dtype=wp.transform),
     body_q_prev: wp.array(dtype=wp.transform),
     body_qd: wp.array(dtype=wp.spatial_vector),
@@ -1354,7 +1308,7 @@ def apply_conservative_bound_truncation(
 
 
 @wp.kernel
-def VBD_solve_trimesh_no_self_contact(
+def solve_trimesh_no_self_contact(
     dt: float,
     particle_ids_in_color: wp.array(dtype=wp.int32),
     prev_pos: wp.array(dtype=wp.vec3),
@@ -1373,16 +1327,8 @@ def VBD_solve_trimesh_no_self_contact(
     edge_bending_properties: wp.array(dtype=float, ndim=2),
     adjacency: ForceElementAdjacencyInfo,
     # contact info
-    soft_contact_ke: float,
-    soft_contact_kd: float,
-    friction_mu: float,
-    friction_epsilon: float,
     particle_forces: wp.array(dtype=wp.vec3),
     particle_hessians: wp.array(dtype=wp.mat33),
-    # ground-particle contact
-    has_ground: bool,
-    ground: wp.array(dtype=float),
-    particle_radius: wp.array(dtype=float),
     # output
     pos_new: wp.array(dtype=wp.vec3),
 ):
@@ -1472,25 +1418,6 @@ def VBD_solve_trimesh_no_self_contact(
         f = f + f_edge
         h = h + h_edge
 
-    if has_ground:
-        ground_normal = wp.vec3(ground[0], ground[1], ground[2])
-        ground_level = ground[3]
-        ground_contact_force, ground_contact_hessian = evaluate_ground_contact_force_hessian(
-            particle_pos,
-            particle_prev_pos,
-            particle_radius[particle_index],
-            ground_normal,
-            ground_level,
-            soft_contact_ke,
-            soft_contact_kd,
-            friction_mu,
-            friction_epsilon,
-            dt,
-        )
-
-        f = f + ground_contact_force
-        h = h + ground_contact_hessian
-
     h = h + particle_hessians[particle_index]
     f = f + particle_forces[particle_index]
 
@@ -1500,7 +1427,7 @@ def VBD_solve_trimesh_no_self_contact(
 
 
 @wp.kernel
-def VBD_copy_particle_positions_back(
+def copy_particle_positions_back(
     particle_ids_in_color: wp.array(dtype=wp.int32),
     pos: wp.array(dtype=wp.vec3),
     pos_new: wp.array(dtype=wp.vec3),
@@ -1544,7 +1471,7 @@ def convert_body_particle_contact_data_kernel(
 
 
 @wp.kernel
-def VBD_accumulate_contact_force_and_hessian(
+def accumulate_contact_force_and_hessian(
     # inputs
     dt: float,
     current_color: int,
@@ -1725,7 +1652,7 @@ def VBD_accumulate_contact_force_and_hessian(
 
 
 @wp.kernel
-def VBD_accumulate_contact_force_and_hessian_no_self_contact(
+def accumulate_contact_force_and_hessian_no_self_contact(
     # inputs
     dt: float,
     current_color: int,
@@ -1812,14 +1739,6 @@ def VBD_solve_trimesh_with_self_contact_penetration_free(
     particle_hessians: wp.array(dtype=wp.mat33),
     pos_prev_collision_detection: wp.array(dtype=wp.vec3),
     particle_conservative_bounds: wp.array(dtype=float),
-    # ground-particle contact
-    has_ground: bool,
-    ground: wp.array(dtype=float),
-    soft_contact_ke: float,
-    soft_contact_kd: float,
-    friction_mu: float,
-    friction_epsilon: float,
-    particle_radius: wp.array(dtype=float),
     # output
     pos_new: wp.array(dtype=wp.vec3),
 ):
@@ -1900,25 +1819,6 @@ def VBD_solve_trimesh_with_self_contact_penetration_free(
             f = f + f_edge
             h = h + h_edge
 
-    if has_ground:
-        ground_normal = wp.vec3(ground[0], ground[1], ground[2])
-        ground_level = ground[3]
-        ground_contact_force, ground_contact_hessian = evaluate_ground_contact_force_hessian(
-            particle_pos,
-            particle_prev_pos,
-            particle_radius[particle_index],
-            ground_normal,
-            ground_level,
-            soft_contact_ke,
-            soft_contact_kd,
-            friction_mu,
-            friction_epsilon,
-            dt,
-        )
-
-        f = f + ground_contact_force
-        h = h + ground_contact_hessian
-
     # fmt: off
     if wp.static("overall_force_hessian" in VBD_DEBUG_PRINTING_OPTIONS):
         wp.printf(
@@ -1964,9 +1864,9 @@ class VBDSolver(SolverBase):
         # or you can use your custom coloring
         builder.set_coloring(user_provided_particle_coloring)
 
-        model = modelbuilder.finalize()
+        model = builder.finalize()
 
-        solver = newton.VBDSolver(model)
+        solver = newton.solvers.VBDSolver(model)
 
         # simulation loop
         for i in range(100):
@@ -1979,6 +1879,8 @@ class VBDSolver(SolverBase):
         model: Model,
         iterations: int = 10,
         handle_self_contact: bool = False,
+        soft_contact_radius: float = 0.2,
+        soft_contact_margin: float = 0.2,
         integrate_with_external_rigid_solver: bool = False,
         penetration_free_conservative_bound_relaxation: float = 0.42,
         friction_epsilon: float = 1e-2,
@@ -1993,6 +1895,10 @@ class VBDSolver(SolverBase):
                 to the `step` function.
             iterations: Number of VBD iterations per step.
             handle_self_contact: whether to self-contact.
+            soft_contact_radius: The radius used for soft contact detection. This is the distance at which particles
+                will start to interact with each other.
+            soft_contact_margin: The margin used for soft contact detection. This is the distance at which particles
+                will be considered in contact generation. It should be larger than `soft_contact_radius` to avoid missing contacts.
             integrate_with_external_rigid_solver: an indicator of coupled rigid body - cloth simulation.  When set to
                 `True`, the solver assumes the rigid body solve is handled  externally.
             penetration_free_conservative_bound_relaxation: Relaxation factor for conservative penetration-free projection.
@@ -2029,12 +1935,15 @@ class VBDSolver(SolverBase):
         self.body_particle_contact_count = wp.zeros((model.particle_count,), dtype=wp.int32, device=self.device)
 
         self.handle_self_contact = handle_self_contact
+        self.soft_contact_radius = soft_contact_radius
+        self.soft_contact_margin = soft_contact_margin
 
+        soft_contact_max = model.shape_count * model.particle_count
         if handle_self_contact:
-            if self.model.soft_contact_margin < self.model.soft_contact_radius:
+            if soft_contact_margin < soft_contact_radius:
                 raise ValueError(
-                    "Model.soft_contact_margin is smaller than Model.soft_contact_radius, this will result in missing contacts and cause instability.\n"
-                    "It is advisable to make Model.soft_contact_margin 1.5~2 times larger than Model.soft_contact_radius."
+                    "soft_contact_margin is smaller than soft_contact_radius, this will result in missing contacts and cause instability.\n"
+                    "It is advisable to make soft_contact_margin 1.5-2 times larger than soft_contact_radius."
                 )
 
             self.conservative_bound_relaxation = penetration_free_conservative_bound_relaxation
@@ -2055,10 +1964,10 @@ class VBDSolver(SolverBase):
             self.collision_evaluation_kernel_launch_size = max(
                 self.model.particle_count * NUM_THREADS_PER_COLLISION_PRIMITIVE,
                 self.model.edge_count * NUM_THREADS_PER_COLLISION_PRIMITIVE,
-                self.model.soft_contact_max,
+                soft_contact_max,
             )
         else:
-            self.collision_evaluation_kernel_launch_size = self.model.soft_contact_max
+            self.collision_evaluation_kernel_launch_size = soft_contact_max
 
         # spaces for particle force and hessian
         self.particle_forces = wp.zeros(self.model.particle_count, dtype=wp.vec3, device=self.device)
@@ -2081,81 +1990,80 @@ class VBDSolver(SolverBase):
         adjacency = ForceElementAdjacencyInfo()
         edges_array = model.edge_indices.to("cpu")
 
-        if edges_array.size:
-            # build vertex-edge adjacency data
-            num_vertex_adjacent_edges = wp.zeros(shape=(self.model.particle_count,), dtype=wp.int32, device="cpu")
+        with wp.ScopedDevice("cpu"):
 
+            if edges_array.size:
+                # build vertex-edge adjacency data
+                num_vertex_adjacent_edges = wp.zeros(shape=(self.model.particle_count,), dtype=wp.int32)
+
+                wp.launch(
+                    kernel=self.count_num_adjacent_edges,
+                    inputs=[edges_array, num_vertex_adjacent_edges],
+                    dim=1,
+                )
+
+                num_vertex_adjacent_edges = num_vertex_adjacent_edges.numpy()
+                vertex_adjacent_edges_offsets = np.empty(shape=(self.model.particle_count + 1,), dtype=wp.int32)
+                vertex_adjacent_edges_offsets[1:] = np.cumsum(2 * num_vertex_adjacent_edges)[:]
+                vertex_adjacent_edges_offsets[0] = 0
+                adjacency.v_adj_edges_offsets = wp.array(vertex_adjacent_edges_offsets, dtype=wp.int32)
+
+                # temporal variables to record how much adjacent edges has been filled to each vertex
+                vertex_adjacent_edges_fill_count = wp.zeros(
+                    shape=(self.model.particle_count,), dtype=wp.int32
+                )
+
+                edge_adjacency_array_size = 2 * num_vertex_adjacent_edges.sum()
+                # vertex order: o0: 0, o1: 1, v0: 2, v1: 3,
+                adjacency.v_adj_edges = wp.empty(shape=(edge_adjacency_array_size,), dtype=wp.int32)
+
+                wp.launch(
+                    kernel=self.fill_adjacent_edges,
+                    inputs=[
+                        edges_array,
+                        adjacency.v_adj_edges_offsets,
+                        vertex_adjacent_edges_fill_count,
+                        adjacency.v_adj_edges,
+                    ],
+                    dim=1,
+                )
+            else:
+                adjacency.v_adj_edges_offsets = wp.empty(shape=(0,), dtype=wp.int32)
+                adjacency.v_adj_edges = wp.empty(shape=(0,), dtype=wp.int32)
+
+            # compute adjacent triangles
+
+            # count number of adjacent faces for each vertex
+            face_indices = model.tri_indices.to("cpu")
+            num_vertex_adjacent_faces = wp.zeros(shape=(self.model.particle_count,), dtype=wp.int32)
             wp.launch(
-                kernel=self.count_num_adjacent_edges,
-                inputs=[edges_array, num_vertex_adjacent_edges],
-                dim=1,
-                device="cpu",
+                kernel=self.count_num_adjacent_faces, inputs=[face_indices, num_vertex_adjacent_faces], dim=1
             )
 
-            num_vertex_adjacent_edges = num_vertex_adjacent_edges.numpy()
-            vertex_adjacent_edges_offsets = np.empty(shape=(self.model.particle_count + 1,), dtype=wp.int32)
-            vertex_adjacent_edges_offsets[1:] = np.cumsum(2 * num_vertex_adjacent_edges)[:]
-            vertex_adjacent_edges_offsets[0] = 0
-            adjacency.v_adj_edges_offsets = wp.array(vertex_adjacent_edges_offsets, dtype=wp.int32, device="cpu")
+            # preallocate memory based on counting results
+            num_vertex_adjacent_faces = num_vertex_adjacent_faces.numpy()
+            vertex_adjacent_faces_offsets = np.empty(shape=(self.model.particle_count + 1,), dtype=wp.int32)
+            vertex_adjacent_faces_offsets[1:] = np.cumsum(2 * num_vertex_adjacent_faces)[:]
+            vertex_adjacent_faces_offsets[0] = 0
+            adjacency.v_adj_faces_offsets = wp.array(vertex_adjacent_faces_offsets, dtype=wp.int32)
 
-            # temporal variables to record how much adjacent edges has been filled to each vertex
-            vertex_adjacent_edges_fill_count = wp.zeros(
-                shape=(self.model.particle_count,), dtype=wp.int32, device="cpu"
-            )
+            vertex_adjacent_faces_fill_count = wp.zeros(shape=(self.model.particle_count,), dtype=wp.int32)
 
-            edge_adjacency_array_size = 2 * num_vertex_adjacent_edges.sum()
-            # vertex order: o0: 0, o1: 1, v0: 2, v1: 3,
-            adjacency.v_adj_edges = wp.empty(shape=(edge_adjacency_array_size,), dtype=wp.int32, device="cpu")
+            face_adjacency_array_size = 2 * num_vertex_adjacent_faces.sum()
+            # (face, vertex_order) * num_adj_faces * num_particles
+            # vertex order: v0: 0, v1: 1, o0: 2, v2: 3
+            adjacency.v_adj_faces = wp.empty(shape=(face_adjacency_array_size,), dtype=wp.int32)
 
             wp.launch(
-                kernel=self.fill_adjacent_edges,
+                kernel=self.fill_adjacent_faces,
                 inputs=[
-                    edges_array,
-                    adjacency.v_adj_edges_offsets,
-                    vertex_adjacent_edges_fill_count,
-                    adjacency.v_adj_edges,
+                    face_indices,
+                    adjacency.v_adj_faces_offsets,
+                    vertex_adjacent_faces_fill_count,
+                    adjacency.v_adj_faces,
                 ],
                 dim=1,
-                device="cpu",
             )
-        else:
-            adjacency.v_adj_edges_offsets = wp.empty(shape=(0,), dtype=wp.int32, device="cpu")
-            adjacency.v_adj_edges = wp.empty(shape=(0,), dtype=wp.int32, device="cpu")
-
-        # compute adjacent triangles
-
-        # count number of adjacent faces for each vertex
-        face_indices = model.tri_indices.to("cpu")
-        num_vertex_adjacent_faces = wp.zeros(shape=(self.model.particle_count,), dtype=wp.int32, device="cpu")
-        wp.launch(
-            kernel=self.count_num_adjacent_faces, inputs=[face_indices, num_vertex_adjacent_faces], dim=1, device="cpu"
-        )
-
-        # preallocate memory based on counting results
-        num_vertex_adjacent_faces = num_vertex_adjacent_faces.numpy()
-        vertex_adjacent_faces_offsets = np.empty(shape=(self.model.particle_count + 1,), dtype=wp.int32)
-        vertex_adjacent_faces_offsets[1:] = np.cumsum(2 * num_vertex_adjacent_faces)[:]
-        vertex_adjacent_faces_offsets[0] = 0
-        adjacency.v_adj_faces_offsets = wp.array(vertex_adjacent_faces_offsets, dtype=wp.int32, device="cpu")
-
-        vertex_adjacent_faces_fill_count = wp.zeros(shape=(self.model.particle_count,), dtype=wp.int32, device="cpu")
-
-        face_adjacency_array_size = 2 * num_vertex_adjacent_faces.sum()
-        # (face, vertex_order) * num_adj_faces * num_particles
-        # vertex order: v0: 0, v1: 1, o0: 2, v2: 3
-        adjacency.v_adj_faces = wp.empty(shape=(face_adjacency_array_size,), dtype=wp.int32, device="cpu")
-
-        wp.launch(
-            kernel=self.fill_adjacent_faces,
-            inputs=[
-                face_indices,
-                adjacency.v_adj_faces_offsets,
-                vertex_adjacent_faces_fill_count,
-                adjacency.v_adj_faces,
-            ],
-            dim=1,
-            device="cpu",
-        )
 
         return adjacency
 
@@ -2165,13 +2073,11 @@ class VBDSolver(SolverBase):
             raise ValueError("model must be the one used to initialize VBDSolver")
 
         if self.handle_self_contact:
-            self.simulate_one_step_with_collisions_penetration_free(model, state_in, state_out, dt, control)
+            self.simulate_one_step_with_collisions_penetration_free(model, state_in, state_out, control, contacts, dt)
         else:
-            self.simulate_one_step_no_self_contact(model, state_in, state_out, dt, control)
+            self.simulate_one_step_no_self_contact(model, state_in, state_out, control, contacts, dt)
 
-    def simulate_one_step_no_self_contact(
-        self, model: Model, state_in: State, state_out: State, dt: float, control: Control | None = None
-    ):
+    def simulate_one_step_no_self_contact(self, model: Model, state_in: State, state_out: State, control: Control, contacts: Contacts, dt: float):
         wp.launch(
             kernel=forward_step,
             inputs=[
@@ -2195,7 +2101,7 @@ class VBDSolver(SolverBase):
 
             for color in range(len(self.model.particle_color_groups)):
                 wp.launch(
-                    kernel=VBD_accumulate_contact_force_and_hessian_no_self_contact,
+                    kernel=accumulate_contact_force_and_hessian_no_self_contact,
                     dim=self.collision_evaluation_kernel_launch_size,
                     inputs=[
                         dt,
@@ -2209,26 +2115,26 @@ class VBDSolver(SolverBase):
                         self.model.soft_contact_mu,
                         self.friction_epsilon,
                         self.model.particle_radius,
-                        self.model.soft_contact_particle,
-                        self.model.soft_contact_count,
-                        self.model.soft_contact_max,
+                        contacts.soft_contact_particle,
+                        contacts.soft_contact_count,
+                        contacts.soft_contact_max,
                         self.model.shape_materials,
                         self.model.shape_body,
                         state_out.body_q if self.integrate_with_external_rigid_solver else state_in.body_q,
                         state_in.body_q if self.integrate_with_external_rigid_solver else None,
                         self.model.body_qd,
                         self.model.body_com,
-                        self.model.soft_contact_shape,
-                        self.model.soft_contact_body_pos,
-                        self.model.soft_contact_body_vel,
-                        self.model.soft_contact_normal,
+                        contacts.soft_contact_shape,
+                        contacts.soft_contact_body_pos,
+                        contacts.soft_contact_body_vel,
+                        contacts.soft_contact_normal,
                     ],
                     outputs=[self.particle_forces, self.particle_hessians],
                     device=self.device,
                 )
 
                 wp.launch(
-                    kernel=VBD_solve_trimesh_no_self_contact,
+                    kernel=solve_trimesh_no_self_contact,
                     inputs=[
                         dt,
                         self.model.particle_color_groups[color],
@@ -2247,16 +2153,8 @@ class VBDSolver(SolverBase):
                         self.model.edge_rest_length,
                         self.model.edge_bending_properties,
                         self.adjacency,
-                        self.model.soft_contact_ke,
-                        self.model.soft_contact_kd,
-                        self.model.soft_contact_mu,
-                        self.friction_epsilon,
                         self.particle_forces,
                         self.particle_hessians,
-                        #   ground-particle contact
-                        self.model.ground,
-                        self.model.ground_plane,
-                        self.model.particle_radius,
                     ],
                     outputs=[
                         state_out.particle_q,
@@ -2266,21 +2164,23 @@ class VBDSolver(SolverBase):
                 )
 
                 wp.launch(
-                    kernel=VBD_copy_particle_positions_back,
-                    inputs=[self.model.particle_color_groups[color], state_in.particle_q, state_out.particle_q],
+                    kernel=copy_particle_positions_back,
+                    inputs=[self.model.particle_color_groups[color], state_in.particle_q],
+                    outputs=[state_out.particle_q],
                     dim=self.model.particle_color_groups[color].size,
                     device=self.device,
                 )
 
         wp.launch(
             kernel=update_velocity,
-            inputs=[dt, self.particle_q_prev, state_out.particle_q, state_out.particle_qd],
+            inputs=[dt, self.particle_q_prev, state_out.particle_q],
+            outputs=[state_out.particle_qd],
             dim=self.model.particle_count,
             device=self.device,
         )
 
     def simulate_one_step_with_collisions_penetration_free(
-        self, model: Model, state_in: State, state_out: State, dt: float, control: Control = None
+        self, model: Model, state_in: State, state_out: State, control: Control, contacts: Contacts, dt: float
     ):
         # collision detection before initialization to compute conservative bounds for initialization
         self.collision_detection_penetration_free(state_in, dt)
@@ -2315,45 +2215,46 @@ class VBDSolver(SolverBase):
             self.particle_hessians.zero_()
 
             for color in range(len(self.model.particle_color_groups)):
-                wp.launch(
-                    kernel=VBD_accumulate_contact_force_and_hessian,
-                    dim=self.collision_evaluation_kernel_launch_size,
-                    inputs=[
-                        dt,
-                        color,
-                        self.particle_q_prev,
-                        state_in.particle_q,
-                        self.model.particle_colors,
-                        self.model.tri_indices,
-                        self.model.edge_indices,
-                        # self-contact
-                        self.trimesh_collision_info,
-                        self.model.soft_contact_radius,
-                        self.model.soft_contact_ke,
-                        self.model.soft_contact_kd,
-                        self.model.soft_contact_mu,
-                        self.friction_epsilon,
-                        self.trimesh_collision_detector.edge_edge_parallel_epsilon,
-                        # body-particle contact
-                        self.model.particle_radius,
-                        self.model.soft_contact_particle,
-                        self.model.soft_contact_count,
-                        self.model.soft_contact_max,
-                        self.model.shape_materials,
-                        self.model.shape_body,
-                        state_out.body_q if self.integrate_with_external_rigid_solver else state_in.body_q,
-                        state_in.body_q if self.integrate_with_external_rigid_solver else None,
-                        self.model.body_qd,
-                        self.model.body_com,
-                        self.model.soft_contact_shape,
-                        self.model.soft_contact_body_pos,
-                        self.model.soft_contact_body_vel,
-                        self.model.soft_contact_normal,
-                    ],
-                    outputs=[self.particle_forces, self.particle_hessians],
-                    device=self.device,
-                    max_blocks=self.model.device.sm_count,
-                )
+                if contacts is not None:
+                    wp.launch(
+                        kernel=accumulate_contact_force_and_hessian,
+                        dim=self.collision_evaluation_kernel_launch_size,
+                        inputs=[
+                            dt,
+                            color,
+                            self.particle_q_prev,
+                            state_in.particle_q,
+                            self.model.particle_colors,
+                            self.model.tri_indices,
+                            self.model.edge_indices,
+                            # self-contact
+                            self.trimesh_collision_info,
+                            self.soft_contact_radius,
+                            self.model.soft_contact_ke,
+                            self.model.soft_contact_kd,
+                            self.model.soft_contact_mu,
+                            self.friction_epsilon,
+                            self.trimesh_collision_detector.edge_edge_parallel_epsilon,
+                            # body-particle contact
+                            self.model.particle_radius,
+                            contacts.soft_contact_particle,
+                            contacts.soft_contact_count,
+                            contacts.soft_contact_max,
+                            self.model.shape_materials,
+                            self.model.shape_body,
+                            state_out.body_q if self.integrate_with_external_rigid_solver else state_in.body_q,
+                            state_in.body_q if self.integrate_with_external_rigid_solver else None,
+                            self.model.body_qd,
+                            self.model.body_com,
+                            contacts.soft_contact_shape,
+                            contacts.soft_contact_body_pos,
+                            contacts.soft_contact_body_vel,
+                            contacts.soft_contact_normal,
+                        ],
+                        outputs=[self.particle_forces, self.particle_hessians],
+                        device=self.device,
+                        max_blocks=self.model.device.sm_count,
+                    )
 
                 wp.launch(
                     kernel=VBD_solve_trimesh_with_self_contact_penetration_free,
@@ -2380,13 +2281,6 @@ class VBDSolver(SolverBase):
                         self.particle_hessians,
                         self.pos_prev_collision_detection,
                         self.particle_conservative_bounds,
-                        self.model.ground,
-                        self.model.ground_plane,
-                        self.model.soft_contact_ke,
-                        self.model.soft_contact_kd,
-                        self.model.soft_contact_mu,
-                        self.friction_epsilon,
-                        self.model.particle_radius,
                     ],
                     outputs=[
                         state_out.particle_q,
@@ -2395,7 +2289,7 @@ class VBDSolver(SolverBase):
                 )
 
                 wp.launch(
-                    kernel=VBD_copy_particle_positions_back,
+                    kernel=copy_particle_positions_back,
                     inputs=[self.model.particle_color_groups[color], state_in.particle_q, state_out.particle_q],
                     dim=self.model.particle_color_groups[color].size,
                     device=self.device,
@@ -2410,15 +2304,15 @@ class VBDSolver(SolverBase):
 
     def collision_detection_penetration_free(self, current_state: State, dt: float):
         self.trimesh_collision_detector.refit(current_state.particle_q)
-        self.trimesh_collision_detector.vertex_triangle_collision_detection(self.model.soft_contact_margin)
-        self.trimesh_collision_detector.edge_edge_collision_detection(self.model.soft_contact_margin)
+        self.trimesh_collision_detector.vertex_triangle_collision_detection(self.soft_contact_margin)
+        self.trimesh_collision_detector.edge_edge_collision_detection(self.soft_contact_margin)
 
         self.pos_prev_collision_detection.assign(current_state.particle_q)
         wp.launch(
             kernel=compute_particle_conservative_bound,
             inputs=[
                 self.conservative_bound_relaxation,
-                self.model.soft_contact_margin,
+                self.soft_contact_margin,
                 self.adjacency,
                 self.trimesh_collision_detector.collision_info,
             ],
